backend/requirements.txt
fastapi
uvicorn
httpx
google-generativeai
supabase
python-dotenv